+++ examples/music-player/src/2_main.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { createHashRouter, RouterProvider } from "react-router-dom";
@@ -22,8 +23,9 @@ import { useUploadExampleData } from "./lib/useUploadExampleData";
  *
  * `<Jazz.Provider/>` also runs our account migration
  */
+
 const Jazz = createJazzReactApp({
-    AccountSchema: MusicaAccount
+    AccountSchema: MusicaAccount,
 });
 
 export const { useAccount, useCoState, useAcceptInvite } = Jazz;
@@ -62,7 +64,10 @@ function JazzAndAuth({ children }: { children: React.ReactNode }) {
 
     return (
         <>
-            <Jazz.Provider auth={auth} peer="wss://mesh.jazz.tools/?key=music-player-example-jazz@gcmp.io">
+            <Jazz.Provider
+                auth={auth}
+                peer="wss://mesh.jazz.tools/?key=music-player-example-jazz@gcmp.io"
+            >
                 {children}
             </Jazz.Provider>
             <DemoAuthBasicUI appName="Jazz Music Player" state={state} />