+++ packages/jazz-tools/src/index.web.ts
@@ -1,9 +1,4 @@
-export {
-    cojsonInternals,
-    MAX_RECOMMENDED_TX_SIZE,
-    WasmCrypto,
-    PureJSCrypto,
-} from "cojson";
+export { cojsonInternals, MAX_RECOMMENDED_TX_SIZE, WasmCrypto } from "cojson";
 export type {
     InviteSecret,
     Peer,
@@ -37,5 +32,5 @@ export {
     ephemeralCredentialsAuth,
     AnonymousJazzAgent,
     createAnonymousJazzContext,
-    randomSessionProvider
+    randomSessionProvider,
 } from "./internal.js";