+++ packages/jazz-tools/src/tests/coStream.test.ts
@@ -1,5 +1,5 @@
 import { expect, describe, test } from "vitest";
-import { connectedPeers } from "cojson/src/streamUtils.js";
+import { connectedPeers } from "cojson/src/streamUtils.ts";
 import {
     BinaryCoStream,
     ID,
@@ -11,7 +11,7 @@ import {
     cojsonInternals,
     createJazzContext,
     fixedCredentialsAuth,
-} from "../index.js";
+} from "../index.web.js";
 import { randomSessionProvider } from "../internal.js";
 
 const Crypto = await WasmCrypto.create();