+++ e2e/CoValues/CHANGELOG.md
@@ -1,9 +1,25 @@
 # @jazz-e2e/covalues
 
+## 0.0.84
+
+### Patch Changes
+
+-   Updated dependencies
+    -   hash-slash@0.2.1
+
+## 0.0.83
+
+### Patch Changes
+
+-   Updated dependencies
+    -   cojson@0.8.3
+    -   jazz-react@0.8.3
+    -   jazz-tools@0.8.3
+
 ## 0.0.82
 
 ### Patch Changes
 
-- Updated dependencies [a075f90]
-  - jazz-tools@0.8.2
-  - jazz-react@0.8.2
+-   Updated dependencies [a075f90]
+    -   jazz-tools@0.8.2
+    -   jazz-react@0.8.2