+++ examples/chat-clerk/CHANGELOG.md
@@ -1,5 +1,22 @@
 # jazz-example-chat
 
+## 0.0.84
+
+### Patch Changes
+
+- Updated dependencies
+  - hash-slash@0.2.1
+
+## 0.0.83
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson@0.8.3
+  - jazz-react@0.8.3
+  - jazz-react-auth-clerk@0.8.3
+  - jazz-tools@0.8.3
+
 ## 0.0.82
 
 ### Patch Changes