+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,21 @@
 # jazz-run
 
+## 0.8.4
+
+### Patch Changes
+
+-   561924f: fix broken import from jazz-run
+
+## 0.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   cojson@0.8.3
+    -   cojson-storage-sqlite@0.8.3
+    -   cojson-transport-ws@0.8.3
+    -   jazz-tools@0.8.3
+
 ## 0.8.2
 
 ### Patch Changes